backend/requirements.txt
python-dotenv
psycopg2
sqlalchemy
pyexcel
pyexcel-io
pyexcel-xls
pyexcel-xlsx
emails
bcrypt
passlib
PyJWT
pyOpenSSL
thrift
flask
flask-cors
greenlet
gevent
werkzeug